Add Toggle All button

+++ src/main/hypernet/handler/ButtonHandlerFactory.java
@@ -7,6 +7,7 @@ import hypernet.handler.button.Disable;
 import hypernet.handler.button.Dummy;
 import hypernet.handler.button.Enable;
 import hypernet.handler.button.RefreshAll;
+import hypernet.handler.button.ToggleAll;
 import hypernet.panel.ControlRow;
 import hypernet.panel.QueryRow;
 
@@ -16,13 +17,10 @@ public class ButtonHandlerFactory {
         if (buttonId == null) {
             return new Dummy();
         }
-
         String buttonType = String.valueOf(buttonId);
-
         if (buttonType.contains("#")) {
             return getIntelHandler(buttonType);
         }
-
         return getQueriesHandler(buttonType);
     }
 
@@ -30,33 +28,28 @@ public class ButtonHandlerFactory {
         String[] buttonIds = b.split("#");
         String buttonType = buttonIds[0];
         int queryNumber = Integer.parseInt(buttonIds[1]);
-
-        switch (buttonType) {
-            case QueryRow.BUTTON_DELETE:
-                return new Delete(queryNumber);
-
-            case QueryRow.BUTTON_DISABLE:
-                return new Disable(queryNumber);
-
-            case QueryRow.BUTTON_ENABLE:
-                return new Enable(queryNumber);
+        ButtonHandler handler = new Dummy();
+        if (buttonType.equals(QueryRow.BUTTON_DELETE)) {
+            handler = new Delete(queryNumber);
+        } else if (buttonType.equals(QueryRow.BUTTON_DISABLE)) {
+            handler = new Disable(queryNumber);
+        } else if (buttonType.equals(QueryRow.BUTTON_ENABLE)) {
+            handler = new Enable(queryNumber);
         }
-
-        return new Dummy();
+        return handler;
     }
 
     private static ButtonHandler getQueriesHandler(String buttonType) {
-        switch (buttonType) {
-            case ControlRow.BUTTON_ADD:
-                return new Add();
-
-            case ControlRow.BUTTON_DELETE_ALL:
-                return new DeleteAll();
-
-            case ControlRow.BUTTON_REFRESH_ALL:
-                return new RefreshAll();
+        ButtonHandler handler = new Dummy();
+        if (buttonType.equals(ControlRow.BUTTON_ADD)) {
+            handler = new Add();
+        } else if (buttonType.equals(ControlRow.BUTTON_DELETE_ALL)) {
+            handler = new DeleteAll();
+        } else if (buttonType.equals(ControlRow.BUTTON_REFRESH_ALL)) {
+            handler = new RefreshAll();
+        } else if (buttonType.equals(ControlRow.BUTTON_TOGGLE_ALL)) {
+            handler = new ToggleAll();
         }
-
-        return new Dummy();
+        return handler;
     }
 }
